tf-m: Update non-secure region definitions
In nRF53 and nRF91 series the security attributes (TrustZone)
for the flash and the RAM regions are configured using the SPU
peripheral. The SPU peripheral adds some hardware limitations regarding
the partitions.
These are:
1) The veneers needs to be placed at the end of a secure partition
2) The size of veneers region is limited by the FLASHNSC[n].SIZE options
3) The flash/RAM regions need to be aligned to the SPU region size

In nRF54L15 the flash and the RAM security attributes are configured
using the SAU and the MPC which have different limitations.
The MPC has smaller region size size of 4K allowing better granularity.
The non-secure callable region doesn't have the limitations 1 and 2.
It only needs to be aligned by 32 bytes as the SAU requires.

This change updates the region_definition header to remove the
non-secure callable limitations for nRF54L15. It also updates
the assertion logic and the relevant documentation.

Ref: NCSDK-25023

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>