sysbuild: pass the dynamic partition to partition manager
Pass the dynamic partition to partition manager when a DOMAIN is passed
to CMake partition_manager() function.

Remove the CPUNET domain restriction as the use of any domain should be
handled. Domain is only passed to the function when it's not the domain
of the main application.

This fixes the issues like:
Traceback (most recent call last):
  File "/<path>/partition_manager_output.py", line 251, in <module>
    main()
  File "/<path>/partition_manager_output.py", line 247, in main
    write_gpm_config(gpm_config, greg_config, name, header_file)
  File "/<path>/partition_manager_output.py", line 142, in
                                                      write_gpm_config
    if any('span' in x for x in gpm_config[domain][image]):
                                ~~~~~~~~~~~~~~~~~~^^^^^^^
KeyError: 'empty_app_core'

as the dynamic partition is now passed to the partition manager script
in all cases where domain is specified.

Jira: NCSDK-23805

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>