mpsl: cx: Rename MPSL_CX_THREAD to MPSL_CX_3WIRE and fix dependency
MPSL_CX_THREAD is a confusing name for a generic 3-wire coex
implementation that is compatible with all protocols that support
generic coexistence, including Bluetooth.
Rename this to MPSL_CX_3WIRE, update the documentation, and remove the
dependency on Bluetooth not being enabled in KConfig.
Update the documentation to not mention a direction pin, as this does
not match the implementation.

Signed-off-by: Aleksandar Stanoev <aleksandar.stanoev@nordicsemi.no>