partition_manager_output.py: Convert affiliation string to list
In the documentation we have to following example:

```yaml
      fatfs_storage:
          affiliation: disk
          extra_params: {
              disk_name: "SD",
              disk_cache_size: 4096,
              disk_sector_size: 512,
              disk_read_only: 0
          }
          placement:
              before: [end]
              align: {start: 4096}
          inside: [nonsecure_storage]
          size: 65536
```

Unfortunately this does not compile in NCS this is due to the
affiliation parameter is assumed to be a list in the python code.
Resulting in the parsed parameter to be `['d', 'i', 's', 'k']` instead
of `['disk']`.

To generate the correct parameter the correct example would be:

```yaml
      fatfs_storage:
          affiliation:
            - disk
          extra_params: {
              disk_name: "SD",
              disk_cache_size: 4096,
              disk_sector_size: 512,
              disk_read_only: 0
          }
          placement:
              before: [end]
              align: {start: 4096}
          inside: [nonsecure_storage]
          size: 65536
```

However since I think it should be supported to just provide 1
affiliation as a string I've added support code in python to convert the
affiliation parameter into a single item list if it's provided as a
string.

Ref. NCSIDB-1228

Signed-off-by: Sigvart Hovland <sigvart.hovland@nordicsemi.no>