samples: 802154_sniffer: reduce chances of missed frames
I observed that with the default number of RX buffers, 16,
the sniffer starts dropping frames when sending 40 frames
of size 115B, with the gap of 1ms (tested with "ltx 40 1"
PHY test sample's command).

Increase the number of RX buffers the reduce the frame drops.
With the new setting, I am able to run "ltx 250 1" without
a drop.

Signed-off-by: Damian Krolik <damian.krolik@nordicsemi.no>