doc: build internal documentation docset when available
Internal documentation is hosted in a private repository. When
the documentation is built for internal hosting (available only to
Nordic-internal audience), the documentation build process must
get access to the internal repository and filter-in the documentation
repository in the west tool. The tool puts the repository in the
repo folder structure. If the folder containing the internal
documentation is available, build scripts add the "internal" docset
to the documentation build process. This way the documentation
containing this additional docset can be hosted for internal usage.

Signed-off-by: Hubert Miś <hubert.mis@nordicsemi.no>