applications: serial_lte_modem: ppp: fix state machine when PPP goes down
When receiving NET_EVENT_PPP_PHASE_DEAD used to check that
ppp_state != PPP_STATE_STOPPED to decide whether PPP was running.

This provoked undesired PPP autorestarts when the link was going down,
because the NET_EVENT_PPP_PHASE_DEAD event was received while PPP was
going down, in which case ppp_state == PPP_STATE_STOPPING.

The proper way is to check ppp_state == PPP_STATE_RUNNING, because
we only want to restart PPP if the link was not being closed down
when receiving the event.

Signed-off-by: Tomi Fontanilles <tomi.fontanilles@nordicsemi.no>