partition_manager: HW model 1 fix for non TF-M socs
The workaround that we have for the hardware model v1 with
TF-M seems to cause build issues with non TF-M devices because
of the missing NRF_SPU_* Kconfigs. Set a default 0x0 to the
non TF-M devices to avoid the build issue. This configuration
is not used for non TF-M devices, so it doesn't have any
negative sideffects.

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>