net: openthread: Introduce OpenThread build report for NCS
The OpenThread report contains information about the current
NCS and OpenThread revisions, Thread feature set, and the
difference between the current and the latest nrfxlib OpenThread
directory to detect the potential changes that may impact
the Thread work.

The report is shown as a build console log and as a .txt file
build artefact.

Described the OpenThread build report and how to use it in the
certification process.
Updated release notes.

Signed-off-by: Arkadiusz Balys <arkadiusz.balys@nordicsemi.no>