nfc: Do not enable EGU on nRF54H
The definition of NFC_LOW_LATENCY_IRQ has been updated to
depend on !SOC_SERIES_NRF54HX as that was the case before
the EGU instance was added to device tree.
We should find a better way of selecting
SWIs. Currently it is already a bit hacky for 54L with

bool "SWI0"
depends on HAS_HW_NRF_EGU0 || HAS_HW_NRF_SWI0 || SOC_SERIES_NRF54LX

Signed-off-by: Rubin Gerritsen <rubin.gerritsen@nordicsemi.no>