sdfw_services: extmem: make write and erase optional
To reduce memory utilization, the write and erase operations can now be
optionally enabled. Write and erase operations are needed only when the
SUIT manifest requests the device to write a component to the external
flash.

To upgrade firmware that has a candidate stored in external flash
into the MRAM, read operations are enough.

Signed-off-by: Rafał Kuźnia <rafal.kuznia@nordicsemi.no>