samples: matter: Fixed DFU using internal flash for nRF54L15
The DFU using internal flash for nRF54L15DK does not work properly,
because the mcuboot secondary partition offset is not aligned
to the 4k. Due to that mcuboot erase operation is not performed
properly and trailer information are not removed, what leads
to swapping back the old image.

The problems was solved by increasing the settings size, to move
all partitions and in result achieve aligned mcuboot partitions
offset.

Signed-off-by: Kamil Kasperczyk <kamil.kasperczyk@nordicsemi.no>