samples: suit: Refactor of external flash configurations
The extflash file suffix did not allow to see what configuration
options are needed for enabling specific external flash configurations.

It also caused the need of many duplications and did not really simplify
things.

Another thing which this commit does is renaming the configurations
to emphasize the features added by each on of them.

Signed-off-by: Artur Hadasz <artur.hadasz@nordicsemi.no>